+++ src/object/mod.rs
@@ -1,5 +1,6 @@
 //! General functionality over Ruby objects.
 
+use std::fmt;
 use crate::{
     prelude::*,
     ruby,
@@ -26,7 +27,12 @@ pub use self::{
 ///
 /// All types that implement this trait _must_ be light wrappers around an
 /// [`AnyObject`](struct.AnyObject.html) and thus have the same size and layout.
-pub unsafe trait Object: Copy + Into<AnyObject> + AsRef<AnyObject> + PartialEq<AnyObject> {
+pub unsafe trait Object: Copy
+    + Into<AnyObject>
+    + AsRef<AnyObject>
+    + PartialEq<AnyObject>
+    + fmt::Debug
+{
     /// Returns a unique identifier for an object type to facilitate casting.
     ///
     /// # Safety